added header dependency generation in makefile

+++ cpp-00/ex01/main.cpp
@@ -6,7 +6,7 @@
 /*   By: okoca <okoca@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/20 18:53:00 by okoca             #+#    #+#             */
-/*   Updated: 2024/07/21 21:20:37 by okoca            ###   ########.fr       */
+/*   Updated: 2024/07/22 08:21:22 by okoca            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int main()
 	std::cout << "[ADD] [SEARCH] [EXIT]" << std::endl;
 	while (std::cout << "PhoneBook3000> " << std::flush && std::getline(std::cin, input))
 	{
-		for (int i = 0; i < input.length(); i++)
+		for (size_t i = 0; i < input.length(); i++)
 			input[i] = toupper(input[i]);
 		if (input == "EXIT")
 			break ;
